getting started with pyre, isolating pydantic types as they generate a lot of [13] initialization errors
