+++ .gitignore
@@ -2,5 +2,6 @@
 .env
 static
 venv
+.pyre/
 __pycache__
