+++ Pipfile
@@ -19,6 +19,7 @@ yoyo-migrations = "~=8.1.0"
 pyclean = "~=2.6.0"
 requests = "~=2.28.2"
 scrapy = "~=2.8.0"
+pyre-check = "*"
 
 [requires]
 python_version = "~=3.10.6"
